+++ plugins/loginguard/smsapi/classes/Exception/HostException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SMSApi\Exception;
+
+class HostException extends SmsapiException
+{
+
+}