+++ plugins/loginguard/smsapi/classes/Exception/ProxyException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SMSApi\Exception;
+
+class ProxyException extends SmsapiException
+{
+
+}